# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TraderMemos API
  version: 1.0.0
  description: |
    Self-hosted trading journal API.

    Authenticate with either:
    - A short-lived session JWT from `POST /api/v1/auth/login` (web app)
    - A personal access token (`tm_pat_…`) from Settings → API or `POST /api/v1/access-tokens`

    Send the secret as `Authorization: Bearer <token>`.
servers:
  - url: /
    description: This API host
...
paths:
...
  /api/v1/accounts:
    get:
      tags: [Accounts]
      summary: List accounts
      operationId: listAccounts
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags: [Accounts]
      summary: Create account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccount"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/accounts/{id}:
    get:
      tags: [Accounts]
      summary: Get account
      operationId: getAccount
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [Accounts]
      summary: Update account
      operationId: updateAccount
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                broker:
                  type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      tags: [Accounts]
      summary: Delete account
      operationId: deleteAccount
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
...
  /api/v1/imports:
    get:
      tags: [Import]
      summary: List import batches
      operationId: listImports
      responses:
        "200":
          description: Import batches
    post:
      tags: [Import]
      summary: Preview import (multipart CSV/JSON)
      operationId: previewImport
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Preview
  /api/v1/imports/commit:
    post:
      tags: [Import]
      summary: Commit import without a preview batch (creates account from JSON when needed)
      operationId: commitImportFresh
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                account_id:
                  type: string
      responses:
        "200":
          description: Committed
  /api/v1/imports/{id}/commit:
    post:
      tags: [Import]
      summary: Commit import batch
      operationId: commitImport
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Committed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Session JWT access token **or** personal access token (`tm_pat_…`).
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
...
    Account:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        broker:
          type: string
        account_type:
          type: string
        base_currency:
          type: string
        starting_balance:
          type: number
    CreateAccount:
      type: object
      required: [name]
      properties:
        name:
          type: string
        broker:
          type: string
        account_type:
          type: string
        base_currency:
          type: string
        starting_balance:
          type: number
security:
  - bearerAuth: []